app/core/command_handler.py
# app/core/command_handler.py

import asyncio
from typing import Dict, Any, List
from datetime import datetime
import logging

from app.api.models import AlarmNotification, AlarmSeverity, MachineCommand
from .clients.opcua_client import OPCUAClient

class CommandHandler:
    def __init__(self, opcua_client: OPCUAClient):
        self.opcua_client = opcua_client
        self.logger = logging.getLogger(__name__)
        self.active_alarms: Dict[str, AlarmNotification] = {}

    async def execute_command(self, command: MachineCommand, params: Dict[str, Any] = None) -> bool:
        # Implement your command execution logic here
        pass

    async def monitor_alarms(self):
        # Implement alarm monitoring logic
        pass

    def get_active_alarms(self) -> List[AlarmNotification]:
        return list(self.active_alarms.values())

    async def acknowledge_alarm(self, alarm_code: str) -> bool:
        if alarm_code in self.active_alarms:
            self.active_alarms[alarm_code].acknowledged = True
            return True
        return False
